# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lego Brick Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BrickStateColor = typeof BrickStateColor[keyof typeof BrickStateColor];


export const BrickStateColor = {
  red: 'red',
  blue: 'blue',
} as const;

export interface BrickState {
  color: BrickStateColor;
  /** Name of the current brick holder */
  holder: string;
  updatedAt: string;
}

export interface TransferInput {
  /** Name of the person to transfer the brick to */
  to: string;
}

export type TransferColor = typeof TransferColor[keyof typeof TransferColor];


export const TransferColor = {
  red: 'red',
  blue: 'blue',
} as const;

export interface Transfer {
  id: number;
  color: TransferColor;
  fromHolder: string;
  toHolder: string;
  transferredAt: string;
}

export interface ErrorResponse {
  error: string;
}