# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lego Brick Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bricks/{color}/transfer:
    post:
      operationId: transferBrick
      tags: [bricks]
      summary: Transfer a brick to someone
      parameters:
        - name: color
          in: path
          required: true
          schema:
            type: string
            enum: [red, blue]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "200":
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrickState"
        "400":
          description: Invalid transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    BrickState:
      type: object
      required: [color, holder, updatedAt]
      properties:
        color:
          type: string
          enum: [red, blue]
        holder:
          type: string
          description: Name of the current brick holder
        updatedAt:
          type: string
          format: date-time
    TransferInput:
      type: object
      required: [to]
      properties:
        to:
          type: string
          description: Name of the person to transfer the brick to
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string