# lib/api-zod/src/generated/types/brickState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lego Brick Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { BrickStateColor } from './brickStateColor';

export interface BrickState {
  color: BrickStateColor;
  /** Name of the current brick holder */
  holder: string;
  updatedAt: Date;
}