# lib/api-zod/src/generated/types/transferInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lego Brick Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TransferInput {
  /** Name of the person to transfer the brick to */
  to: string;
}